+++ Nos/Views/Onboarding/OnboardingView.swift
@@ -9,22 +9,33 @@ import SwiftUI
 import Dependencies
 import Logger
 
+class OnboardingState: ObservableObject {
+    @Published var flow: OnboardingFlow = .createAccount
+    @Published var step: OnboardingStep = .onboardingStart {
+        didSet {
+            path.append(step)
+        }
+    }
+    @Published var path = NavigationPath()
+}
+
+enum OnboardingFlow {
+    case createAccount
+    case loginToExistingAccount
+}
+
+enum OnboardingStep {
+    case onboardingStart
+    case ageVerification
+    case notOldEnough
+    case termsOfService
+    case finishOnboarding
+}
+
 struct OnboardingView: View {
     @Environment(\.managedObjectContext) private var viewContext
 
-    enum OnboardingStep {
-        case onboardingStart
-        case addPrivateKey
-        case ageVerification
-        case notOldEnough
-        case termsOfService
-        case createAccount
-    }
-    
-    enum OnboardingFlow {
-        case createAccount
-        case loginToExistingAccount
-    }
+    @ObservedObject var state = OnboardingState()
     
     /// Completion to be called when all onboarding steps are complete
     let completion: () -> Void
@@ -41,221 +52,36 @@ struct OnboardingView: View {
         }
     }
     
-    @State var privateKeyString = ""
-    
-    @State var showError = false
-    
     @State var flow: OnboardingFlow = .createAccount
-    
-    @State var path = NavigationPath()
     
     @Dependency(\.analytics) private var analytics
     
-    var loginView: some View {
-        VStack {
-            Form {
-                Section {
-                    TextField("NSec1", text: $privateKeyString)
-                        .foregroundColor(.textColor)
-                } header: {
-                    Localized.pasteYourSecretKey.view
-                        .foregroundColor(.textColor)
-                        .fontWeight(.heavy)
-                }
-                .listRowBackground(LinearGradient(
-                    colors: [Color.cardBgTop, Color.cardBgBottom],
-                    startPoint: .top,
-                    endPoint: .bottom
-                ))
-            }
-            if !privateKeyString.isEmpty {
-                BigActionButton(title: .login) {
-                    if let keyPair = KeyPair(nsec: privateKeyString) {
-                        self.keyPair = keyPair
-                        analytics.identify(with: keyPair)
-                        analytics.importedKey()
-
-                        // Use these to sync
-                        for address in Relay.allKnown {
-                            do {
-                                let relay = try Relay(
-                                    context: viewContext,
-                                    address: address,
-                                    author: CurrentUser.shared.author
-                                )
-                                CurrentUser.shared.onboardingRelays.append(relay)
-                            } catch {
-                                Log.error(error.localizedDescription)
-                            }
-                        }
-                        try? CurrentUser.shared.context.save()
-
-                        completion()
-                    } else {
-                        self.keyPair = nil
-                        self.showError = true
-                    }
-                }
-                .padding(.horizontal, 24)
-            }
-        }
-        .scrollContentBackground(.hidden)
-        .background(Color.appBg)
-        .navigationTitle(Localized.loginToYourAccount.string)
-        .alert(isPresented: $showError) {
-            Alert(
-                title: Localized.invalidKey.view,
-                message: Localized.couldNotReadPrivateKeyMessage.view
-            )
-        }
-    }
-    
     var body: some View {
-        TabView(selection: $selectedTab) {
-            VStack {
-                Image.nosLogo
-                    .resizable()
-                    .frame(width: 235.45, height: 67.1)
-                    .padding(.top, 155)
-                    .padding(.bottom, 10)
-                PlainText(Localized.onboardingTitle.string)
-                    .font(.custom("ClarityCity-Bold", size: 25.21))
-                    .fontWeight(.heavy)
-                    .foregroundStyle(
-                        LinearGradient(
-                            colors: [
-                                Color(hex: "#F08508"),
-                                Color(hex: "#F43F75")
-                            ],
-                            startPoint: .bottomLeading,
-                            endPoint: .topTrailing
-                        )
-                        .blendMode(.normal)
-                    )
-                Spacer()
-                BigActionButton(title: .createAccount) {
-                    flow = .createAccount
-                    selectedTab = .ageVerification
-                }
-                .padding()
-                Button(Localized.loginToYourAccount.string) {
-                    flow = .loginToExistingAccount
-                    selectedTab = .ageVerification
-                }
-            }
-            .background(Color.appBg)
-            .tag(OnboardingStep.onboardingStart)
-            
-            // Age verification
-            VStack {
-                PlainText(Localized.ageVerificationTitle.string)
-                    .multilineTextAlignment(.center)
-                    .padding(.top, 92)
-                    .padding(.bottom, 20)
-                    .padding(.horizontal, 77.5)
-                    .font(.custom("ClarityCity-Bold", size: 34, relativeTo: .largeTitle))
-                    .foregroundStyle(
-                        LinearGradient(
-                            colors: [
-                                Color(hex: "#F08508"),
-                                Color(hex: "#F43F75")
-                            ],
-                            startPoint: .bottomLeading,
-                            endPoint: .topTrailing
-                        )
-                        .blendMode(.normal)
-                    )
-                Text(Localized.ageVerificationSubtitle.string)
-                    .foregroundColor(.secondaryTxt)
-                    .multilineTextAlignment(.center)
-                    .padding(.horizontal, 44.5)
-                Spacer()
-                HStack {
-                    BigActionButton(title: .no) {
-                        selectedTab = .notOldEnough
-                    }
-                    BigActionButton(title: .yes) {
-                        selectedTab = .termsOfService
-                    }
-                }
-                .padding(.horizontal, 24)
-            }
-            .background(Color.appBg)
-            .tag(OnboardingStep.ageVerification)
-            
-            // Not old enough
-            VStack {
-                PlainText(Localized.notOldEnoughTitle.string)
-                    .font(.custom("ClarityCity-Bold", size: 34, relativeTo: .largeTitle))
-                    .foregroundStyle(
-                        LinearGradient(
-                            colors: [
-                                Color(hex: "#F08508"),
-                                Color(hex: "#F43F75")
-                            ],
-                            startPoint: .bottomLeading,
-                            endPoint: .topTrailing
-                        )
-                        .blendMode(.normal)
-                    )
-                    .multilineTextAlignment(.center)
-                    .padding(.top, 92)
-                    .padding(.bottom, 20)
-                    .padding(.horizontal, 45)
-                Text(Localized.notOldEnoughSubtitle.string)
-                    .foregroundColor(.secondaryTxt)
-                    .multilineTextAlignment(.center)
-                    .padding(.horizontal, 45)
-                Spacer()
-                BigActionButton(title: .notOldEnoughButton) {
-                    selectedTab = .onboardingStart
-                }
-                .padding(.horizontal, 24)
-            }
-            .background(Color.appBg)
-            .tag(OnboardingStep.notOldEnough)
-            
-            // Terms of Service
-            NavigationStack(path: $path) {
-                VStack {
-                    PlainText(Localized.termsOfServiceTitle.string)
-                        .font(.custom("ClarityCity-Bold", size: 34, relativeTo: .largeTitle))
-                        .foregroundStyle(
-                            LinearGradient(
-                                colors: [
-                                    Color(hex: "#F08508"),
-                                    Color(hex: "#F43F75")
-                                ],
-                                startPoint: .bottomLeading,
-                                endPoint: .topTrailing
-                            )
-                            .blendMode(.normal)
-                        )
-                        .padding(.top, 92)
-                        .padding(.bottom, 60)
-                    ScrollView {
-                        Text(Localized.termsOfService.string)
-                            .foregroundColor(.secondaryTxt)
-                        Rectangle().fill(Color.clear)
-                            .frame(height: 100)
-                    }
-                    .mask(
-                        VStack(spacing: 0) {
-                            Rectangle().fill(Color.black)
-                            LinearGradient(
-                                colors: [Color.black, Color.black.opacity(0)],
-                                startPoint: .top,
-                                endPoint: .bottom
-                            )
-                        }
-                    )
-                    .padding(.horizontal, 44.5)
-                    HStack {
-                        BigActionButton(title: Localized.reject) {
-                            selectedTab = .onboardingStart
-                        }
-                        BigActionButton(title: Localized.accept) {
-                            if flow == .createAccount {
+        NavigationStack(path: $state.path) {
+            OnboardingStartView()
+                .environmentObject(state)
+                .navigationDestination(for: OnboardingStep.self) { step in
+                    switch step {
+                    case .onboardingStart:
+                        OnboardingStartView()
+                            .environmentObject(state)
+                    case .ageVerification:
+                        OnboardingAgeVerificationView()
+                            .environmentObject(state)
+                    case .notOldEnough:
+                        OnboardingNotOldEnoughView()
+                            .environmentObject(state)
+                    case .termsOfService:
+                        OnboardingTermsOfServiceView()
+                            .environmentObject(state)
+                    case .finishOnboarding:
+                        switch state.flow {
+                        case .createAccount:
+                            // hack to allow us to do business logic here... we won't need this once
+                            // we have a dedicated CreateAccountView that handles setting up the current user
+                            // swiftlint: disable redundant_discardable_let
+                            let _ = {
+                            // swiftlint: enable redundant_discardable_let
                                 let keyPair = KeyPair()!
                                 self.keyPair = keyPair
                                 analytics.identify(with: keyPair)
@@ -272,26 +98,15 @@ struct OnboardingView: View {
                                 try? CurrentUser.shared.context.save()
                                 
                                 CurrentUser.shared.publishContactList(tags: [])
-                            }
-                            path.append(flow)
+                            }()
+                            ProfileEditView(author: CurrentUser.shared.author!, createAccountCompletion: completion)
+                        case .loginToExistingAccount:
+                            OnboardingLoginView(completion: completion)
                         }
                     }
-                    .padding(.horizontal, 24)
                 }
-                .navigationDestination(for: OnboardingFlow.self) { flow in
-                    if flow == .loginToExistingAccount {
-                        loginView
-                    } else {
-                        ProfileEditView(author: CurrentUser.shared.author!, createAccountCompletion: completion)
-                    }
-                }
-                .background(Color.appBg)
-            }
-            .tag(OnboardingStep.termsOfService)
         }
     }
-    
-    @State var finishOnboardingAction: Int? = 0
 }
 
 struct OnboardingView_Previews: PreviewProvider {